backend/requirements.txt
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0
pydantic==2.6.4
requests==2.31.0
sentence-transformers==2.6.1
faiss-cpu==1.8.0
pypdf==4.1.0
numpy==1.26.4